requirements.txt
snowflake-connector-python
python-dotenv
streamlit
jupyter
